Fix behaviour when submitting non locking pages twice
Issue #2 describes an issue where if you have your plugin set up in a
non locking manner, and submit the same page to multiple assignments,
then only the latest link will work.

The reason for this is the Mahara generates a token in an mt parameter
that changes on every submit. The non locking behaviour allows multiple
submissions, but to do that, it submits then releases, hence generating
a new token.

This patch, along with a patch in Mahara, allows sending a lock
parameter to Mahara, and then appending the url with the moodle
assignment and the mahara view/collection. Mahara will then send a
request back to Moodle to test whether you have permission to view that
view, and if so, stores the result in the session. This allows all
submission urls to work.

For upgrade paths, as this changes the get_view_url method, no change to
existing submissions is needed.

This patch makes a requirement on either a 4 line core Moodle patch, or
a local plugin, as there are issues with subplugin MNet publishers.