Changes on code review
1. Have updated Mahara so that it signals that it respected
the "lock=false" param, by making the "accesskey" be null. So
I changed the corresponding code on the Moodle side to look
for that.

2. Renamed "get_view_url" to "mnetify_url" to make it clearer
what it does. Removed the code that generates the "gradebook"
URL with "&mnetviewid=X&assignment=X" and put it in view(),
since that's the only place that does that.

3. Updated the comments on the status constants to indicate
their dual usage in new & old Mahara interfaces. Updated status
usage so that if we never lock a Mahara page, we just leave
the status on "STATUS_SELECTED".